refactor: move chat component to bot root component
